Added `setup.py`, `py.typed`, and move the collection of tests to the `__init__.py` from `tests/`
