+++ setup.py
@@ -1,0 +1,9 @@
+from setuptools import setup
+
+setup(
+    name="Proppy",
+    author="Eugen Bronasco",
+    version="0.1",
+    package_data={"proppy": ["py.typed"]},
+    packages=["proppy"]
+)
