+++ tox.ini
@@ -15,10 +15,7 @@ commands = mypy proppy
         flake8 proppy
         pylint proppy
 
-#        flake8-errmsg proppy/*.py
-
 #flake8 #==6.1.0
-#    flake8_errmsg #==0.5.0
 #    mypy #==1.7.0
 #    mypy-extensions #==1.0.0
 #    pydantic #==2.5.2
@@ -26,4 +23,3 @@ commands = mypy proppy
 #    pydash #==7.0.6
 #    pylint #==3.0.2
 #    pytest #==7.4.3-rrequirements.txt
-#
